requirements_fixed.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
bit==0.8.0
requests==2.31.0
Werkzeug==2.3.7
SQLAlchemy>=2.1.0
gunicorn==21.2.0
pytest==7.4.0
